BETA: Fixing minor typos: 	* MagicSkillLaunched should not sent for Toggle skills. 	* Recharge shots to determinate which one blessed or not spiritshots player have. 	* Experience should not be casted as integer. 		* Reported by: angkor_tm

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/model/actor/stat/PcStat.java
@@ -204,7 +204,7 @@ public class PcStat extends PlayableStat
 			else
 			{
 				sm = SystemMessage.getSystemMessage(SystemMessageId.YOU_EARNED_S1_EXP_AND_S2_SP);
-				sm.addLong((int) addToExp);
+				sm.addLong(addToExp);
 				sm.addInt(addToSp);
 			}
 		}
